Working for 64 and 32-bit.
