+++ opencv2.framework/Resources
@@ -1,1 +1,1 @@
-Versions/Current/Resources
+Versions/A/Resources
