+++ opencv2.framework/Versions/A/Headers/core/cuda_devptrs.hpp
@@ -22,7 +22,7 @@
 //
 //   * Redistribution's in binary form must reproduce the above copyright notice,
 //     this list of conditions and the following disclaimer in the documentation
-//     and/or other materials provided with the distribution.
+//     and/or other GpuMaterials provided with the distribution.
 //
 //   * The name of the copyright holders may not be used to endorse or promote products
 //     derived from this software without specific prior written permission.
@@ -177,20 +177,6 @@ namespace cv
 //#undef __CV_GPU_DEPR_BEFORE__
 //#undef __CV_GPU_DEPR_AFTER__
 
-        namespace device
-        {
-            using cv::gpu::PtrSz;
-            using cv::gpu::PtrStep;
-            using cv::gpu::PtrStepSz;
-
-            using cv::gpu::PtrStepSzb;
-            using cv::gpu::PtrStepSzf;
-            using cv::gpu::PtrStepSzi;
-
-            using cv::gpu::PtrStepb;
-            using cv::gpu::PtrStepf;
-            using cv::gpu::PtrStepi;
-        }
     }
 }
