+++ opencv2.framework/Versions/A/Headers/core/internal.hpp
@@ -50,9 +50,6 @@
 
 #include <vector>
 
-#include "opencv2/core/core.hpp"
-#include "opencv2/core/types_c.h"
-
 #if defined WIN32 || defined _WIN32
 #  ifndef WIN32
 #    define WIN32
@@ -254,10 +251,6 @@ namespace cv
         body(range);
     }
 #endif
-
-    // Returns a static string if there is a parallel framework,
-    // NULL otherwise.
-    CV_EXPORTS const char* currentParallelFramework();
 } //namespace cv
 
 #define CV_INIT_ALGORITHM(classname, algname, memberinit) \
@@ -758,9 +751,7 @@ typedef struct CvBigFuncTable
     (tab).fn_2d[CV_64F] = (void*)FUNCNAME##_64f##FLAG
 
 #ifdef __cplusplus
-
-// < Deprecated
-
+//! OpenGL extension table
 class CV_EXPORTS CvOpenGlFuncTab
 {
 public:
@@ -786,16 +777,10 @@ CV_EXPORTS void icvSetOpenGlFuncTab(const CvOpenGlFuncTab* tab);
 
 CV_EXPORTS bool icvCheckGlError(const char* file, const int line, const char* func = "");
 
-// >
-
-namespace cv { namespace ogl {
-CV_EXPORTS bool checkError(const char* file, const int line, const char* func = "");
-}}
-
 #if defined(__GNUC__)
-    #define CV_CheckGlError() CV_DbgAssert( (cv::ogl::checkError(__FILE__, __LINE__, __func__)) )
+    #define CV_CheckGlError() CV_DbgAssert( (::icvCheckGlError(__FILE__, __LINE__, __func__)) )
 #else
-    #define CV_CheckGlError() CV_DbgAssert( (cv::ogl::checkError(__FILE__, __LINE__)) )
+    #define CV_CheckGlError() CV_DbgAssert( (::icvCheckGlError(__FILE__, __LINE__)) )
 #endif
 
 #endif //__cplusplus
