+++ opencv2.framework/Versions/A/Headers/core/opengl_interop.hpp
@@ -22,7 +22,7 @@
 //
 //   * Redistribution's in binary form must reproduce the above copyright notice,
 //     this list of conditions and the following disclaimer in the documentation
-//     and/or other materials provided with the distribution.
+//     and/or other GpuMaterials provided with the distribution.
 //
 //   * The name of the copyright holders may not be used to endorse or promote products
 //     derived from this software without specific prior written permission.
@@ -46,238 +46,289 @@
 #ifdef __cplusplus
 
 #include "opencv2/core/core.hpp"
-#include "opencv2/core/opengl_interop_deprecated.hpp"
 
-namespace cv { namespace ogl {
-
-/////////////////// OpenGL Objects ///////////////////
-
+namespace cv
+{
 //! Smart pointer for OpenGL buffer memory with reference counting.
-class CV_EXPORTS Buffer
+class CV_EXPORTS GlBuffer
 {
 public:
-    enum Target
+    enum Usage
     {
-        ARRAY_BUFFER         = 0x8892, //!< The buffer will be used as a source for vertex data
-        ELEMENT_ARRAY_BUFFER = 0x8893, //!< The buffer will be used for indices (in glDrawElements, for example)
-        PIXEL_PACK_BUFFER    = 0x88EB, //!< The buffer will be used for reading from OpenGL textures
-        PIXEL_UNPACK_BUFFER  = 0x88EC  //!< The buffer will be used for writing to OpenGL textures
-    };
-
-    enum Access
-    {
-        READ_ONLY  = 0x88B8,
-        WRITE_ONLY = 0x88B9,
-        READ_WRITE = 0x88BA
+        ARRAY_BUFFER = 0x8892,  // buffer will use for OpenGL arrays (vertices, colors, normals, etc)
+        TEXTURE_BUFFER = 0x88EC // buffer will ise for OpenGL textures
     };
 
     //! create empty buffer
-    Buffer();
-
-    //! create buffer from existed buffer id
-    Buffer(int arows, int acols, int atype, unsigned int abufId, bool autoRelease = false);
-    Buffer(Size asize, int atype, unsigned int abufId, bool autoRelease = false);
+    explicit GlBuffer(Usage usage);
 
     //! create buffer
-    Buffer(int arows, int acols, int atype, Target target = ARRAY_BUFFER, bool autoRelease = false);
-    Buffer(Size asize, int atype, Target target = ARRAY_BUFFER, bool autoRelease = false);
+    GlBuffer(int rows, int cols, int type, Usage usage);
+    GlBuffer(Size size, int type, Usage usage);
 
     //! copy from host/device memory
-    explicit Buffer(InputArray arr, Target target = ARRAY_BUFFER, bool autoRelease = false);
+    GlBuffer(InputArray mat, Usage usage);
 
-    //! create buffer
-    void create(int arows, int acols, int atype, Target target = ARRAY_BUFFER, bool autoRelease = false);
-    void create(Size asize, int atype, Target target = ARRAY_BUFFER, bool autoRelease = false) { create(asize.height, asize.width, atype, target, autoRelease); }
+    void create(int rows, int cols, int type, Usage usage);
+    void create(Size size, int type, Usage usage);
+    void create(int rows, int cols, int type);
+    void create(Size size, int type);
 
-    //! release memory and delete buffer object
     void release();
 
-    //! set auto release mode (if true, release will be called in object's destructor)
-    void setAutoRelease(bool flag);
-
     //! copy from host/device memory
-    void copyFrom(InputArray arr, Target target = ARRAY_BUFFER, bool autoRelease = false);
+    void copyFrom(InputArray mat);
 
-    //! copy to host/device memory
-    void copyTo(OutputArray arr, Target target = ARRAY_BUFFER, bool autoRelease = false) const;
-
-    //! create copy of current buffer
-    Buffer clone(Target target = ARRAY_BUFFER, bool autoRelease = false) const;
-
-    //! bind buffer for specified target
-    void bind(Target target) const;
-
-    //! unbind any buffers from specified target
-    static void unbind(Target target);
+    void bind() const;
+    void unbind() const;
 
     //! map to host memory
-    Mat mapHost(Access access);
+    Mat mapHost();
     void unmapHost();
 
     //! map to device memory
     gpu::GpuMat mapDevice();
     void unmapDevice();
 
-    int rows() const { return rows_; }
-    int cols() const { return cols_; }
-    Size size() const { return Size(cols_, rows_); }
-    bool empty() const { return rows_ == 0 || cols_ == 0; }
+    inline int rows() const { return rows_; }
+    inline int cols() const { return cols_; }
+    inline Size size() const { return Size(cols_, rows_); }
+    inline bool empty() const { return rows_ == 0 || cols_ == 0; }
 
-    int type() const { return type_; }
-    int depth() const { return CV_MAT_DEPTH(type_); }
-    int channels() const { return CV_MAT_CN(type_); }
-    int elemSize() const { return CV_ELEM_SIZE(type_); }
-    int elemSize1() const { return CV_ELEM_SIZE1(type_); }
+    inline int type() const { return type_; }
+    inline int depth() const { return CV_MAT_DEPTH(type_); }
+    inline int channels() const { return CV_MAT_CN(type_); }
+    inline int elemSize() const { return CV_ELEM_SIZE(type_); }
+    inline int elemSize1() const { return CV_ELEM_SIZE1(type_); }
 
-    unsigned int bufId() const;
+    inline Usage usage() const { return usage_; }
 
     class Impl;
-
 private:
-    Ptr<Impl> impl_;
     int rows_;
     int cols_;
     int type_;
+    Usage usage_;
+
+    Ptr<Impl> impl_;
 };
 
-//! Smart pointer for OpenGL 2D texture memory with reference counting.
-class CV_EXPORTS Texture2D
+template <> CV_EXPORTS void Ptr<GlBuffer::Impl>::delete_obj();
+
+//! Smart pointer for OpenGL 2d texture memory with reference counting.
+class CV_EXPORTS GlTexture
 {
 public:
-    enum Format
-    {
-        NONE            = 0,
-        DEPTH_COMPONENT = 0x1902, //!< Depth
-        RGB             = 0x1907, //!< Red, Green, Blue
-        RGBA            = 0x1908  //!< Red, Green, Blue, Alpha
-    };
-
     //! create empty texture
-    Texture2D();
-
-    //! create texture from existed texture id
-    Texture2D(int arows, int acols, Format aformat, unsigned int atexId, bool autoRelease = false);
-    Texture2D(Size asize, Format aformat, unsigned int atexId, bool autoRelease = false);
+    GlTexture();
 
     //! create texture
-    Texture2D(int arows, int acols, Format aformat, bool autoRelease = false);
-    Texture2D(Size asize, Format aformat, bool autoRelease = false);
+    GlTexture(int rows, int cols, int type);
+    GlTexture(Size size, int type);
 
     //! copy from host/device memory
-    explicit Texture2D(InputArray arr, bool autoRelease = false);
+    explicit GlTexture(InputArray mat, bool bgra = true);
 
-    //! create texture
-    void create(int arows, int acols, Format aformat, bool autoRelease = false);
-    void create(Size asize, Format aformat, bool autoRelease = false) { create(asize.height, asize.width, aformat, autoRelease); }
-
-    //! release memory and delete texture object
+    void create(int rows, int cols, int type);
+    void create(Size size, int type);
     void release();
 
-    //! set auto release mode (if true, release will be called in object's destructor)
-    void setAutoRelease(bool flag);
-
     //! copy from host/device memory
-    void copyFrom(InputArray arr, bool autoRelease = false);
+    void copyFrom(InputArray mat, bool bgra = true);
 
-    //! copy to host/device memory
-    void copyTo(OutputArray arr, int ddepth = CV_32F, bool autoRelease = false) const;
-
-    //! bind texture to current active texture unit for GL_TEXTURE_2D target
     void bind() const;
+    void unbind() const;
 
-    int rows() const { return rows_; }
-    int cols() const { return cols_; }
-    Size size() const { return Size(cols_, rows_); }
-    bool empty() const { return rows_ == 0 || cols_ == 0; }
+    inline int rows() const { return rows_; }
+    inline int cols() const { return cols_; }
+    inline Size size() const { return Size(cols_, rows_); }
+    inline bool empty() const { return rows_ == 0 || cols_ == 0; }
 
-    Format format() const { return format_; }
-
-    unsigned int texId() const;
+    inline int type() const { return type_; }
+    inline int depth() const { return CV_MAT_DEPTH(type_); }
+    inline int channels() const { return CV_MAT_CN(type_); }
+    inline int elemSize() const { return CV_ELEM_SIZE(type_); }
+    inline int elemSize1() const { return CV_ELEM_SIZE1(type_); }
 
     class Impl;
-
 private:
-    Ptr<Impl> impl_;
     int rows_;
     int cols_;
-    Format format_;
+    int type_;
+
+    Ptr<Impl> impl_;
+    GlBuffer buf_;
 };
+
+template <> CV_EXPORTS void Ptr<GlTexture::Impl>::delete_obj();
 
 //! OpenGL Arrays
-class CV_EXPORTS Arrays
+class CV_EXPORTS GlArrays
 {
 public:
-    Arrays();
+    inline GlArrays()
+        : vertex_(GlBuffer::ARRAY_BUFFER), color_(GlBuffer::ARRAY_BUFFER), bgra_(true), normal_(GlBuffer::ARRAY_BUFFER), texCoord_(GlBuffer::ARRAY_BUFFER)
+    {
+    }
 
     void setVertexArray(InputArray vertex);
-    void resetVertexArray();
+    inline void resetVertexArray() { vertex_.release(); }
 
-    void setColorArray(InputArray color);
-    void resetColorArray();
+    void setColorArray(InputArray color, bool bgra = true);
+    inline void resetColorArray() { color_.release(); }
 
     void setNormalArray(InputArray normal);
-    void resetNormalArray();
+    inline void resetNormalArray() { normal_.release(); }
 
     void setTexCoordArray(InputArray texCoord);
-    void resetTexCoordArray();
-
-    void release();
-
-    void setAutoRelease(bool flag);
+    inline void resetTexCoordArray() { texCoord_.release(); }
 
     void bind() const;
+    void unbind() const;
 
-    int size() const { return size_; }
-    bool empty() const { return size_ == 0; }
+    inline int rows() const { return vertex_.rows(); }
+    inline int cols() const { return vertex_.cols(); }
+    inline Size size() const { return vertex_.size(); }
+    inline bool empty() const { return vertex_.empty(); }
 
 private:
-    int size_;
-    Buffer vertex_;
-    Buffer color_;
-    Buffer normal_;
-    Buffer texCoord_;
+    GlBuffer vertex_;
+    GlBuffer color_;
+    bool bgra_;
+    GlBuffer normal_;
+    GlBuffer texCoord_;
 };
 
-/////////////////// Render Functions ///////////////////
+//! OpenGL Font
+class CV_EXPORTS GlFont
+{
+public:
+    enum Weight
+    {
+        WEIGHT_LIGHT    = 300,
+        WEIGHT_NORMAL   = 400,
+        WEIGHT_SEMIBOLD = 600,
+        WEIGHT_BOLD     = 700,
+        WEIGHT_BLACK    = 900
+    };
+
+    enum Style
+    {
+        STYLE_NORMAL    = 0,
+        STYLE_ITALIC    = 1,
+        STYLE_UNDERLINE = 2
+    };
+
+    static Ptr<GlFont> get(const std::string& family, int height = 12, Weight weight = WEIGHT_NORMAL, Style style = STYLE_NORMAL);
+
+    void draw(const char* str, int len) const;
+
+    inline const std::string& family() const { return family_; }
+    inline int height() const { return height_; }
+    inline Weight weight() const { return weight_; }
+    inline Style style() const { return style_; }
+
+private:
+    GlFont(const std::string& family, int height, Weight weight, Style style);
+
+    std::string family_;
+    int height_;
+    Weight weight_;
+    Style style_;
+
+    unsigned int base_;
+
+    GlFont(const GlFont&);
+    GlFont& operator =(const GlFont&);
+};
+
+//! render functions
 
 //! render texture rectangle in window
-CV_EXPORTS void render(const Texture2D& tex,
+CV_EXPORTS void render(const GlTexture& tex,
     Rect_<double> wndRect = Rect_<double>(0.0, 0.0, 1.0, 1.0),
     Rect_<double> texRect = Rect_<double>(0.0, 0.0, 1.0, 1.0));
 
 //! render mode
-enum {
-    POINTS         = 0x0000,
-    LINES          = 0x0001,
-    LINE_LOOP      = 0x0002,
-    LINE_STRIP     = 0x0003,
-    TRIANGLES      = 0x0004,
-    TRIANGLE_STRIP = 0x0005,
-    TRIANGLE_FAN   = 0x0006,
-    QUADS          = 0x0007,
-    QUAD_STRIP     = 0x0008,
-    POLYGON        = 0x0009
-};
+namespace RenderMode {
+    enum {
+        POINTS         = 0x0000,
+        LINES          = 0x0001,
+        LINE_LOOP      = 0x0002,
+        LINE_STRIP     = 0x0003,
+        TRIANGLES      = 0x0004,
+        TRIANGLE_STRIP = 0x0005,
+        TRIANGLE_FAN   = 0x0006,
+        QUADS          = 0x0007,
+        QUAD_STRIP     = 0x0008,
+        POLYGON        = 0x0009
+    };
+}
 
 //! render OpenGL arrays
-CV_EXPORTS void render(const Arrays& arr, int mode = POINTS, Scalar color = Scalar::all(255));
-CV_EXPORTS void render(const Arrays& arr, InputArray indices, int mode = POINTS, Scalar color = Scalar::all(255));
+CV_EXPORTS void render(const GlArrays& arr, int mode = RenderMode::POINTS, Scalar color = Scalar::all(255));
 
-}} // namespace cv::gl
+CV_EXPORTS void render(const std::string& str, const Ptr<GlFont>& font, Scalar color, Point2d pos);
 
-namespace cv { namespace gpu {
+//! OpenGL camera
+class CV_EXPORTS GlCamera
+{
+public:
+    GlCamera();
 
-//! set a CUDA device to use OpenGL interoperability
-CV_EXPORTS void setGlDevice(int device = 0);
+    void lookAt(Point3d eye, Point3d center, Point3d up);
+    void setCameraPos(Point3d pos, double yaw, double pitch, double roll);
 
-}}
+    void setScale(Point3d scale);
 
-namespace cv {
+    void setProjectionMatrix(const Mat& projectionMatrix, bool transpose = true);
+    void setPerspectiveProjection(double fov, double aspect, double zNear, double zFar);
+    void setOrthoProjection(double left, double right, double bottom, double top, double zNear, double zFar);
 
-template <> CV_EXPORTS void Ptr<cv::ogl::Buffer::Impl>::delete_obj();
-template <> CV_EXPORTS void Ptr<cv::ogl::Texture2D::Impl>::delete_obj();
+    void setupProjectionMatrix() const;
+    void setupModelViewMatrix() const;
 
+private:
+    Point3d eye_;
+    Point3d center_;
+    Point3d up_;
+
+    Point3d pos_;
+    double yaw_;
+    double pitch_;
+    double roll_;
+
+    bool useLookAtParams_;
+
+    Point3d scale_;
+
+    Mat projectionMatrix_;
+
+    double fov_;
+    double aspect_;
+
+    double left_;
+    double right_;
+    double bottom_;
+    double top_;
+
+    double zNear_;
+    double zFar_;
+
+    bool perspectiveProjection_;
+};
+
+inline void GlBuffer::create(Size _size, int _type, Usage _usage) { create(_size.height, _size.width, _type, _usage); }
+inline void GlBuffer::create(int _rows, int _cols, int _type) { create(_rows, _cols, _type, usage()); }
+inline void GlBuffer::create(Size _size, int _type) { create(_size.height, _size.width, _type, usage()); }
+inline void GlTexture::create(Size _size, int _type) { create(_size.height, _size.width, _type); }
+
+namespace gpu
+{
+    //! set a CUDA device to use OpenGL interoperability
+    CV_EXPORTS void setGlDevice(int device = 0);
 }
+} // namespace cv
 
 #endif // __cplusplus
