+++ opencv2.framework/Versions/A/Headers/core/types_c.h
@@ -315,8 +315,6 @@ CV_INLINE  int  cvRound( double value )
         fistp t;
     }
     return t;
-#elif defined _MSC_VER && defined _M_ARM && defined HAVE_TEGRA_OPTIMIZATION
-    TEGRA_ROUND(value);
 #elif defined HAVE_LRINT || defined CV_ICC || defined __GNUC__
 #  ifdef HAVE_TEGRA_OPTIMIZATION
     TEGRA_ROUND(value);
@@ -324,12 +322,8 @@ CV_INLINE  int  cvRound( double value )
     return (int)lrint(value);
 #  endif
 #else
-    double intpart, fractpart;
-    fractpart = modf(value, &intpart);
-    if ((fabs(fractpart) != 0.5) || ((((int)intpart) % 2) != 0))
-        return (int)(value + (value >= 0 ? 0.5 : -0.5));
-    else
-        return (int)intpart;
+    // while this is not IEEE754-compliant rounding, it's usually a good enough approximation
+    return (int)(value + (value >= 0 ? 0.5 : -0.5));
 #endif
 }
 
@@ -348,8 +342,9 @@ CV_INLINE  int  cvFloor( double value )
     return i - (i > value);
 #else
     int i = cvRound(value);
-    float diff = (float)(value - i);
-    return i - (diff < 0);
+    Cv32suf diff;
+    diff.f = (float)(value - i);
+    return i - (diff.i < 0);
 #endif
 }
 
@@ -365,8 +360,9 @@ CV_INLINE  int  cvCeil( double value )
     return i + (i < value);
 #else
     int i = cvRound(value);
-    float diff = (float)(i - value);
-    return i + (diff < 0);
+    Cv32suf diff;
+    diff.f = (float)(i - value);
+    return i + (diff.i < 0);
 #endif
 }
 
@@ -375,19 +371,31 @@ CV_INLINE  int  cvCeil( double value )
 
 CV_INLINE int cvIsNaN( double value )
 {
+#if 1/*defined _MSC_VER || defined __BORLANDC__
+    return _isnan(value);
+#elif defined __GNUC__
+    return isnan(value);
+#else*/
     Cv64suf ieee754;
     ieee754.f = value;
     return ((unsigned)(ieee754.u >> 32) & 0x7fffffff) +
            ((unsigned)ieee754.u != 0) > 0x7ff00000;
+#endif
 }
 
 
 CV_INLINE int cvIsInf( double value )
 {
+#if 1/*defined _MSC_VER || defined __BORLANDC__
+    return !_finite(value);
+#elif defined __GNUC__
+    return isinf(value);
+#else*/
     Cv64suf ieee754;
     ieee754.f = value;
     return ((unsigned)(ieee754.u >> 32) & 0x7fffffff) == 0x7ff00000 &&
            (unsigned)ieee754.u == 0;
+#endif
 }
