+++ opencv2.framework/Versions/A/Headers/core/version.hpp
@@ -47,23 +47,12 @@
 #ifndef __OPENCV_VERSION_HPP__
 #define __OPENCV_VERSION_HPP__
 
-#define CV_VERSION_EPOCH    2
-#define CV_VERSION_MAJOR    4
-#define CV_VERSION_MINOR    6
-#define CV_VERSION_REVISION 0
+#define CV_MAJOR_VERSION    2
+#define CV_MINOR_VERSION    4
+#define CV_SUBMINOR_VERSION 3
 
 #define CVAUX_STR_EXP(__A)  #__A
 #define CVAUX_STR(__A)      CVAUX_STR_EXP(__A)
-
-#if CV_VERSION_REVISION
-#  define CV_VERSION        CVAUX_STR(CV_VERSION_EPOCH) "." CVAUX_STR(CV_VERSION_MAJOR) "." CVAUX_STR(CV_VERSION_MINOR) "." CVAUX_STR(CV_VERSION_REVISION)
-#else
-#  define CV_VERSION        CVAUX_STR(CV_VERSION_EPOCH) "." CVAUX_STR(CV_VERSION_MAJOR) "." CVAUX_STR(CV_VERSION_MINOR)
-#endif
-
-/* old  style version constants*/
-#define CV_MAJOR_VERSION    CV_VERSION_EPOCH
-#define CV_MINOR_VERSION    CV_VERSION_MAJOR
-#define CV_SUBMINOR_VERSION CV_VERSION_MINOR
+#define CV_VERSION          CVAUX_STR(CV_MAJOR_VERSION) "." CVAUX_STR(CV_MINOR_VERSION) "." CVAUX_STR(CV_SUBMINOR_VERSION)
 
 #endif
