+++ opencv2.framework/Versions/A/Headers/flann/dist.h
@@ -456,6 +456,7 @@ struct Hamming
     ResultType operator()(Iterator1 a, Iterator2 b, size_t size, ResultType /*worst_dist*/ = -1) const
     {
         ResultType result = 0;
+#ifdef __GNUC__
 #ifdef __ARM_NEON__
         {
             uint32x4_t bits = vmovq_n_u32(0);
@@ -472,7 +473,7 @@ struct Hamming
             result = vgetq_lane_s32 (vreinterpretq_s32_u64(bitSet2),0);
             result += vgetq_lane_s32 (vreinterpretq_s32_u64(bitSet2),2);
         }
-#elif __GNUC__
+#else
         {
             //for portability just use unsigned long -- and use the __builtin_popcountll (see docs for __builtin_popcountll)
             typedef unsigned long long pop_t;
@@ -492,8 +493,8 @@ struct Hamming
                 result += __builtin_popcountll(a_final ^ b_final);
             }
         }
-#else // NO NEON and NOT GNUC
-        typedef unsigned long long pop_t;
+#endif //NEON
+#else
         HammingLUT lut;
         result = lut(reinterpret_cast<const unsigned char*> (a),
                      reinterpret_cast<const unsigned char*> (b), size * sizeof(pop_t));
