+++ opencv2.framework/Versions/A/Headers/imgproc/types_c.h
@@ -309,22 +309,8 @@ enum
     // alpha premultiplication
     CV_RGBA2mRGBA = 125,
     CV_mRGBA2RGBA = 126,
-    
-    CV_RGB2YUV_I420 = 127,
-    CV_BGR2YUV_I420 = 128,
-    CV_RGB2YUV_IYUV = CV_RGB2YUV_I420,
-    CV_BGR2YUV_IYUV = CV_BGR2YUV_I420,
 
-    CV_RGBA2YUV_I420 = 129,
-    CV_BGRA2YUV_I420 = 130,
-    CV_RGBA2YUV_IYUV = CV_RGBA2YUV_I420,
-    CV_BGRA2YUV_IYUV = CV_BGRA2YUV_I420,
-    CV_RGB2YUV_YV12  = 131,
-    CV_BGR2YUV_YV12  = 132,
-    CV_RGBA2YUV_YV12 = 133,
-    CV_BGRA2YUV_YV12 = 134,
-
-    CV_COLORCVT_MAX  = 135
+    CV_COLORCVT_MAX  = 127
 };
