+++ opencv2.framework/Versions/A/Headers/stitching/detail/warpers.hpp
@@ -460,7 +460,7 @@ struct SphericalPortraitProjector : ProjectorBase
 
 // Projects image onto unit sphere with origin at (0, 0, 0).
 // Poles are located NOT at (0, -1, 0) and (0, 1, 0) points, BUT at (1, 0, 0) and (-1, 0, 0) points.
-class CV_EXPORTS SphericalPortraitWarper : public RotationWarperBase<SphericalPortraitProjector>
+class SphericalPortraitWarper : public RotationWarperBase<SphericalPortraitProjector>
 {
 public:
     SphericalPortraitWarper(float scale) { projector_.scale = scale; }
@@ -476,7 +476,7 @@ struct CylindricalPortraitProjector : ProjectorBase
 };
 
 
-class CV_EXPORTS CylindricalPortraitWarper : public RotationWarperBase<CylindricalPortraitProjector>
+class CylindricalPortraitWarper : public RotationWarperBase<CylindricalPortraitProjector>
 {
 public:
     CylindricalPortraitWarper(float scale) { projector_.scale = scale; }
@@ -495,7 +495,7 @@ struct PlanePortraitProjector : ProjectorBase
 };
 
 
-class CV_EXPORTS PlanePortraitWarper : public RotationWarperBase<PlanePortraitProjector>
+class PlanePortraitWarper : public RotationWarperBase<PlanePortraitProjector>
 {
 public:
     PlanePortraitWarper(float scale) { projector_.scale = scale; }
