+++ opencv2.framework/Versions/A/Resources/Info.plist
@@ -5,7 +5,7 @@
     <key>CFBundleName</key>
     <string>OpenCV</string>
     <key>CFBundleIdentifier</key>
-    <string>org.opencv</string>
+    <string>com.itseez.opencv</string>
     <key>CFBundleVersion</key>
     <string>2.4</string>
     <key>CFBundleShortVersionString</key>
