+++ opencv2.framework/opencv2
@@ -1,1 +1,1 @@
-Versions/Current/opencv2
+Versions/A/opencv2
